refactor(common) : Relocation files
