+++ src/main/java/com/multi/travel/auth/jwt/service/JwtFilter.java
@@ -1,4 +1,4 @@
-package com.multi.travel.auth.jwt;
+package com.multi.travel.auth.jwt.service;
 
 
 import com.fasterxml.jackson.core.JsonProcessingException;
