+++ src/main/java/com/multi/travel/auth/jwt/service/TokenProvider.java
@@ -1,4 +1,4 @@
-package com.multi.travel.auth.jwt;
+package com.multi.travel.auth.jwt.service;
 
 
 import com.multi.travel.auth.user.CustomUser;
